MINOR: Simplify member handling in CppClass
Updated CppClass class to collect usings, structs, constants and enums
placed inside into a generic "members" collection instead of four
separate ones. This is similar to what CppFile class already does.

There is no need to distinguish between the elements of these four
types. Their order is determined by TopologicalSort dependency sorting,
not by their exact type.

+++ transpiler/src/test/java/com/here/ivi/api/generator/baseapi/TopologicalSortTest.java
@@ -13,15 +13,7 @@ package com.here.ivi.api.generator.baseapi;
 
 import static org.junit.Assert.assertEquals;
 
-import com.here.ivi.api.model.cppmodel.CppComplexTypeRef;
-import com.here.ivi.api.model.cppmodel.CppConstant;
-import com.here.ivi.api.model.cppmodel.CppElement;
-import com.here.ivi.api.model.cppmodel.CppEnum;
-import com.here.ivi.api.model.cppmodel.CppField;
-import com.here.ivi.api.model.cppmodel.CppStruct;
-import com.here.ivi.api.model.cppmodel.CppTypeDefRef;
-import com.here.ivi.api.model.cppmodel.CppUsing;
-import com.here.ivi.api.model.cppmodel.CppValue;
+import com.here.ivi.api.model.cppmodel.*;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
@@ -188,7 +180,7 @@ public class TopologicalSortTest {
     assertEquals(elements.size(), sortedElements.size());
 
     for (int i = 0; i < elements.size(); ++i) {
-      int index = expectedOrder.get(i).intValue();
+      int index = expectedOrder.get(i);
       assertEquals(elements.get(index), sortedElements.get(i));
     }
   }
